refactor: group CUDA communication modules
