+++ lmdeploy/pytorch/engine/executor/ray_executor.py
@@ -11,7 +11,7 @@ from ray.util.placement_group import PlacementGroup
 from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 
 from lmdeploy.pytorch import envs as _envs
-from lmdeploy.pytorch.backends.cuda.communicator import should_try_symm_mem
+from lmdeploy.pytorch.backends.cuda.comm.communicator import should_try_symm_mem
 from lmdeploy.pytorch.backends.selector import init_backend
 from lmdeploy.pytorch.config import BackendConfig, CacheConfig, DistConfig, MiscConfig, ModelConfig, SpecDecodeConfig
 from lmdeploy.pytorch.devices import DeviceContext, get_device_manager
